update user type student

+++ database/migrations/2021_01_25_040141_add_columns_users_table.php
@@ -14,10 +14,10 @@ class AddColumnsUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->tinyInteger('region')
+            $table->integer('region')
                     ->after('address')
                     ->nullable();
-            $table->tinyInteger('part_time')
+            $table->integer('part_time')
                     ->after('region')
                     ->nullable();
         });
